requirements.txt
streamlit==1.6.0
pandas==1.4.1
numpy==1.22.2
opencv-python==4.5.5.64
Pillow==9.0.1